Refactor: Update all JsDoc to be more consistent.
