+++ src/reporter/Reporter.ts
@@ -17,55 +17,68 @@ export interface ReporterPlugin {
   options?: Record<any, any>;
 }
 
+/**
+ * Configuration options for generating a report.
+ * 
+ * @interface ReportOptions
+ */
+export interface ReportOptions {
+  /**
+   * A map containing report data where:
+   * The key is a unique string identifier for the report
+   * The value is an object of type `PoolResult`
+   * 
+   * @type {Map<string, PoolResult>}
+   */
+  reports: Map<string, PoolResult>;
+
+  /**
+   * The directory where the report should be saved.
+   * This is required if the reporter type is `"file"`. If not provided, the report may be logged to the console.
+   * 
+   * @type {string | undefined}
+   */
+  outputDir?: string;
+
+  /**
+   * Additional configuration options for the report generation.
+   * You can pass custom settings based on the reporter's needs, such as a custom format or logging level.
+   * 
+   * @type {object}
+   */
+  [key: string]: any;
+}
+
+/**
+ * A reporter plugin responsible for generating and outputting reports.
+ * 
+ * @interface Reporter
+ */
 export interface Reporter {
-  /** Name of the reporter plugin */
+  /** Name of the reporter plugin. This is typically a descriptive name, e.g., "FileReporter".
+   * 
+   * @type {string}
+   */
   name: string;
 
   /**
-   * The type of the reporter plugin.
-   *
-   * - `"file"`: Writes the report to a file.
-   * - `"console"`: Outputs the report to the console.
+   * The type of the reporter plugin, which determines how the report is generated and output.
+   * 
+   * @type {('file' | 'console')}
    */
   type: "file" | "console";
 
   /**
    * Generates a report based on the provided options.
-   *
-   * @param {Object} options - Configuration for generating the report.
-   * @param {Map<string, PoolResult>} options.reports - A map containing report data where the key is a string identifier and the value is a `PoolResult`.
-   * @param {string} [options.outputDir] - Directory where the report should be saved. Required if the plugin type is `"file"`.
-   *
-   * @example
-   * const plugin: ReporterPlugin = {
-   *   name: "FileReporter",
-   *   type: "file",
-   *   report: async (options) => {
-   *     console.log("Generating report...");
-   *     // Implementation here...
-   *   },
-   * };
-   *
-   * @returns {Promise<void>} Resolves when the report is successfully generated.
+   * 
+   * @param {ReportOptions} options - Configuration for generating the report.
+   * @returns {Promise<any>} A `Promise` that resolves when the report is successfully generated. If an error occurs, it should be thrown.
+   * @throws {Error} If there is an issue generating the report (e.g., missing required output directory for file-based reports).
+   * 
    */
-  report(options: {
-    /**
-     * A map containing report data where:
-     * - The key is a unique string identifier for the report.
-     * - The value is an object of type `PoolResult`.
-     */
-    reports: Map<string, PoolResult>;
-
-    /**
-     * The directory where the report should be saved.
-     */
-    outputDir?: string;
-
-    [key: string]: any;
-  }): Promise<void>;
+  report(options: ReportOptions): Promise<any>;
 }
 
-// Validation logic for `veve.reporters`
 
 export async function report(
   reports: Map<string, PoolResult>,
