+++ src/utils/xml.ts
@@ -3,76 +3,6 @@
  * Licensed under the MIT License.
  * See the LICENSE file in the project root for license information.
  */
-/**
- * Represents the contract for an XML document writer.
- * Provides methods to build XML documents with support for tags, attributes, and content.
- */
-export interface XML {
-  /**
-   * Opens a new XML tag with optional attributes.
-   *
-   * @param name - The name of the tag
-   * @param attributes - Optional key-value pairs of attributes
-   * @returns The XMLWriter instance for method chaining
-   *
-   * @example
-   * writer.openTag('person', { id: 1, type: 'employee' })
-   */
-  openTag(name: string, attributes?: Record<string, string | number>): XML;
-
-  /**
-   * Adds a self-closing XML tag with optional attributes.
-   *
-   * @param name - The name of the tag
-   * @param attributes - Optional key-value pairs of attributes
-   * @returns The XMLWriter instance for method chaining
-   *
-   * @example
-   * writer.selfClosingTag('image', { src: 'profile.jpg', width: 100 })
-   */
-  selfClosingTag(
-    name: string,
-    attributes?: Record<string, string | number>,
-  ): XML;
-
-  /**
-   * Adds an XML tag with content and optional attributes.
-   *
-   * @param name - The name of the tag
-   * @param content - The text content of the tag
-   * @param attributes - Optional key-value pairs of attributes
-   * @returns The XMLWriter instance for method chaining
-   *
-   * @example
-   * writer.tag('name', 'John Doe', { type: 'full' })
-   */
-  tag(
-    name: string,
-    content: string,
-    attributes?: Record<string, string | number>,
-  ): XML;
-
-  /**
-   * Closes the most recently opened tag.
-   *
-   * @param name - The name of the tag to close
-   * @returns The XMLWriter instance for method chaining
-   *
-   * @example
-   * writer.openTag('person').closeTag('person')
-   */
-  closeTag(name: string): XML;
-
-  /**
-   * Generates the complete XML document as a string.
-   *
-   * @returns The full XML document
-   *
-   * @example
-   * const xmlString = writer.toString();
-   */
-  toString(): string;
-}
 
 export class XML {
   private buffer: string[] = [];
