+++ types/global.d.ts
@@ -1,14 +1,4 @@
-import {
-  Assertion,
-  HookFunction,
-  MethodNames,
-  MethodType,
-  Options,
-  TestFunction,
-  TestReport,
-  TrackedFunction,
-  TrackFn,
-} from ".";
+import { HookFunction, Options, TestFunction, TestReport } from ".";
 
 declare global {
   /**
@@ -191,106 +181,6 @@ declare global {
    * @example const report = await testCase.run();
    */
   function run(): Promise<TestReport>;
-
-  /**
-   * Creates a new assertion instance for the provided value.
-   * - This function performs assertions that will **throw an error** if the condition fails.
-   * - Use this when you need hard failure handling for invalid conditions.
-   *
-   * @param {any} received - The value to be asserted.
-   * @example
-   * assert(42).toBe(42); // Passes
-   * assert("hello").toBe("world"); // Throws an error
-   * @returns {Assertion} An `Assertion` instance configured to throw errors on failure.
-   */
-  function assert(received: any): Assertion;
-
-  /**
-   * Creates a new assertion instance for the provided value.
-   * - This function performs non-throwing assertions, which do not raise errors on failure.
-   * - Use this for scenarios where you want to test conditions without interrupting execution.
-   *
-   * @param {any} received - The value to be strictly asserted.
-   * @example
-   * is(42).toBe(42); // Passes
-   * is("hello").toBe("world"); // Does not throw, but marks failure
-   * @returns {Assertion} An `Assertion` instance configured to silently handle failures.
-   */
-  function is(received: any): Assertion;
-
-  /**
-   * Creates a tracked version of a given function, preserving its original type signature while
-   * adding tracking capabilities. The returned function maintains the same behavior as the original
-   * while providing additional methods for tracking calls, arguments, and results.
-   *
-   * @template T - The type of the function being tracked, must extend (...args: any[]) => any
-   * @param {T} implementation - The original function implementation to track
-   * @returns {TrackedFunction<T>} A function that combines the original implementation with tracking capabilities
-   *
-   * @example
-   * // Track a simple addition function
-   * const add = (a: number, b: number) => a + b;
-   * const trackedAdd = Fn(add);
-   * trackedAdd(1, 2); // Returns 3
-   * console.log(trackedAdd.getCallCount()); // Returns 1
-   *
-   * @example
-   * // Track an async function
-   * const fetchData = async (id: string) => ({ id, data: 'some data' });
-   * const trackedFetch = Fn(fetchData);
-   * await trackedFetch('123');
-   * console.log(trackedFetch.getAllArgs()); // Returns [['123']]
-   *
-   * @example
-   * // Modify tracked function behavior
-   * const greet = (name: string) => `Hello ${name}`;
-   * const trackedGreet = Fn(greet);
-   * trackedGreet.return('Fixed response');
-   * console.log(trackedGreet('Alice')); // Returns 'Fixed response'
-   */
-  function Fn<T extends (...args: any[]) => any>(
-    implementation: T,
-  ): TrackedFunction<T>;
-  
-  /**
-   * Replaces a method on an object with a tracked version while preserving its original type signature.
-   * The original method is replaced with a tracked version that maintains the same behavior but provides
-   * additional tracking capabilities. The tracked version is both assigned to the object and returned
-   * for convenience.
-   *
-   * @template T - The type of the object containing the method to track
-   * @template K - The key type of the method to track, must be a key of T
-   * @param {T} obj - The object containing the method to track
-   * @param {K & MethodNames<T>} method - The name of the method to track
-   * @returns {TrackedFunction<MethodType<T, K>>} A tracked version of the specified method
-   * @throws {Error} If the specified method doesn't exist on the object or isn't a function
-   *
-   * @example
-   * // Track a method on a simple object
-   * const calculator = {
-   *   add: (a: number, b: number) => a + b
-   * };
-   * const trackedAdd = spyOn(calculator, 'add');
-   * calculator.add(2, 3); // Returns 5
-   * console.log(trackedAdd.getCallCount()); // Returns 1
-   *
-   * @example
-   * // Track and modify method behavior
-   * const api = {
-   *   fetch: async (url: string) => ({ data: 'response' })
-   * };
-   * const trackedFetch = spyOn(api, 'fetch');
-   * trackedFetch.throw(new Error('Network error'));
-   * try {
-   *   await api.fetch('/data');
-   * } catch (error) {
-   *   console.log(error.message); // Prints: Network error
-   * }
-   */
-  function spyOn<T extends object, K extends keyof T>(
-    obj: T,
-    method: K & MethodNames<T>,
-  ): TrackedFunction<MethodType<T, K>>;
 }
 
 export {};
